Add pattern planner.
